src/scaffolder.py
import os

class AgentScaffolder:
    def __init__(self, structure_file, destination):
        self.structure_file = structure_file
        self.destination = destination

    def create_project_structure(self):
        with open(self.structure_file, 'r') as file:
            lines = file.readlines()

        current_path = ""
        for line in lines:
            stripped_line = line.strip()
            indent_level = (len(line) - len(stripped_line)) // 4

            if stripped_line.endswith('/'):
                current_path = os.path.join(self.destination, stripped_line.rstrip('/'))
                os.makedirs(current_path, exist_ok=True)
            else:
                file_path = os.path.join(current_path, stripped_line)
                os.makedirs(os.path.dirname(file_path), exist_ok=True)
                with open(file_path, 'w') as f:
                    f.write(f"# This is a placeholder for {stripped_line}")

        print(f"Project structure created at '{self.destination}'")

if __name__ == "__main__":
    import argparse

    parser = argparse.ArgumentParser(description="Create a project structure.")
    parser.add_argument("structure_file", type=str, help="Path to the text file containing the project structure")
    parser.add_argument("destination", type=str, help="Destination directory")

    args = parser.parse_args()

    creator = AgentScaffolder(args.structure_file, args.destination)
    creator.create_project_structure()
